requirements.txt
aiogram==2.16
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.8.1
async-timeout==4.0.1
attrs==21.2.0
Babel==2.9.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.9
dateparser==1.1.0
dnspython==2.1.0
environs==8.0.0
frozenlist==1.2.0
idna==3.3
marshmallow==3.14.1
mongoengine==0.23.1
multidict==5.2.0
numpy==1.21.4
Pillow==8.4.0
pip==21.1.2
pymongo==3.12.2
python-binance==1.0.15
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz-deprecation-shim==0.1.0.post0
pytz==2021.3
regex==2021.11.10
requests==2.26.0
setuptools==57.0.0
six==1.16.0
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
ujson==4.3.0
urllib3==1.26.7
websockets==9.1
wheel==0.36.2
yarl==1.7.2
